chore: config package refactor
