+++ src/main/java/it/gov/pagopa/pu/processecexutions/util/HttpUtils.java
@@ -1,6 +1,6 @@
 package it.gov.pagopa.pu.processecexutions.util;
 
-import it.gov.pagopa.pu.processecexutions.config.HttpClientConfig;
+import it.gov.pagopa.pu.processecexutions.config.rest.HttpClientConfig;
 import org.apache.hc.client5.http.config.ConnectionConfig;
 import org.apache.hc.client5.http.impl.io.PoolingHttpClientConnectionManagerBuilder;
 import org.apache.hc.client5.http.ssl.TlsSocketStrategy;
